/deletelvl fix on maps with perbuildrank higher than the player's minor mapranking fix

+++ Commands/CmdDeleteLvl.cs
@@ -24,10 +24,8 @@ using MCForge.SQL;
 //using MySql.Data.Types;
 
 
-namespace MCForge.Commands
-{
-    public class CmdDeleteLvl : Command
-    {
+namespace MCForge.Commands {
+    public class CmdDeleteLvl : Command {
         public override string name { get { return "deletelvl"; } }
         public override string shortcut { get { return ""; } }
         public override string type { get { return "mod"; } }
@@ -35,51 +33,60 @@ namespace MCForge.Commands
         public override LevelPermission defaultRank { get { return LevelPermission.Admin; } }
         public CmdDeleteLvl() { }
 
-        public override void Use(Player p, string message)
-        {
-            if (message == "") { Help(p); return; }
+        public override void Use(Player p, string message) {
+            if (message == "" || message.Split().Length > 1) { Help(p); return; }
             Level foundLevel = Level.Find(message);
-            if (foundLevel != null) foundLevel.Unload();
+            if (foundLevel != null) {
+                if (foundLevel.permissionbuild > p.group.Permission) {
+                    Player.SendMessage(p, "%cYou can't delete levels with a perbuild rank higher than yours!");
+                    return;
+                }
+                foundLevel.Unload();
+            }
 
             if (foundLevel == Server.mainLevel) { Player.SendMessage(p, "Cannot delete the main level."); return; }
 
-            try
-            {
+            try {
                 if (!Directory.Exists("levels/deleted")) Directory.CreateDirectory("levels/deleted");
 
-                if (File.Exists("levels/" + message + ".lvl"))
-                {
-                    if (File.Exists("levels/deleted/" + message + ".lvl"))
-                    {
+                if (File.Exists("levels/" + message + ".lvl")) {
+
+                    using (StreamReader reader = new StreamReader("levels/level properties/" + message + ".properties")) {
+                        string line;
+                        while ((line = reader.ReadLine()) != null) {
+                            if (line[0] == '#') continue;
+                            if (line.Split()[0].ToLower() == "perbuild") {
+                                if (Level.PermissionFromName(line.Split()[2].ToLower()) > p.group.Permission) {
+                                    Player.SendMessage(p, "%cYou can't delete levels with a perbuild rank higher than yours!");
+                                    return;
+                                }
+                                break;
+                            }
+                        }
+                    }
+
+                    if (File.Exists("levels/deleted/" + message + ".lvl")) {
                         int currentNum = 0;
                         while (File.Exists("levels/deleted/" + message + currentNum + ".lvl")) currentNum++;
 
                         File.Move("levels/" + message + ".lvl", "levels/deleted/" + message + currentNum + ".lvl");
-                    }
-                    else
-                    {
+                    } else {
                         File.Move("levels/" + message + ".lvl", "levels/deleted/" + message + ".lvl");
                     }
                     Player.SendMessage(p, "Created backup.");
 
-                    try { File.Delete("levels/level properties/" + message + ".properties"); }
-                    catch { }
-                    try { File.Delete("levels/level properties/" + message); }
-                    catch { }
+                    try { File.Delete("levels/level properties/" + message + ".properties"); } catch { }
+                    try { File.Delete("levels/level properties/" + message); } catch { }
 
                     if (Server.useMySQL) MySQL.executeQuery("DROP TABLE `Block" + message + "`, `Portals" + message + "`, `Messages" + message + "`, `Zone" + message + "`"); else SQLite.executeQuery("DROP TABLE Block" + message + ", Portals" + message + ", Messages" + message + " Zone" + message + "");
 
                     Player.GlobalMessage("Level " + message + " was deleted.");
-                }
-                else
-                {
+                } else {
                     Player.SendMessage(p, "Could not find specified level.");
                 }
-            }
-            catch (Exception e) { Player.SendMessage(p, "Error when deleting."); Server.ErrorLog(e); }
+            } catch (Exception e) { Player.SendMessage(p, "Error when deleting."); Server.ErrorLog(e); }
         }
-        public override void Help(Player p)
-        {
+        public override void Help(Player p) {
             Player.SendMessage(p, "/deletelvl [map] - Completely deletes [map] (portals, MBs, everything");
             Player.SendMessage(p, "A backup of the map will be placed in the levels/deleted folder");
         }
